+++ ParallaxGenLib/include/ParallaxGenWarnings.hpp
@@ -2,6 +2,7 @@
 
 #include <mutex>
 #include <unordered_set>
+#include <map>
 
 #include "ParallaxGenDirectory.hpp"
 
@@ -22,11 +23,9 @@ private:
   };
 
   // Trackers for warnings
-  static std::unordered_set<std::pair<std::wstring, std::wstring>, PairHash>
-      MismatchWarnTracker;                    /** Keeps tabs on WARN mismatch messages to avoid duplicates */
+  static std::unordered_map<std::wstring, std::unordered_set<std::wstring>> MismatchWarnTracker;  // matched mod to set of base mods
   static std::mutex MismatchWarnTrackerMutex; /** Mutex for MismatchWarnTracker */
-  static std::unordered_set<std::pair<std::wstring, std::wstring>, PairHash>
-      MismatchWarnDebugTracker;                    /** Keeps tabs on DEBUG mismatch messages to avoid duplicates */
+  static std::unordered_map<std::wstring, std::unordered_set<std::pair<std::wstring,std::wstring>, PairHash>> MismatchWarnDebugTracker;
   static std::mutex MismatchWarnDebugTrackerMutex; /** Mutex for MismatchWarnDebugTracker */
 
   static std::unordered_set<std::pair<std::wstring, std::wstring>, PairHash>
@@ -46,7 +45,7 @@ public:
   static void init(ParallaxGenDirectory *PGD, const std::unordered_map<std::wstring, int> *ModPriority);
 
   /**
-   * @brief Posts warning about a mismatch between diffuse/normal and a matched path. Will not repost the same warning.
+   * @brief store warning about a mismatch between diffuse/normal and a matched path. Will not repost the same warning.
    *
    * @param MatchedPath Path that was matched (for example _m or _p file)
    * @param BaseTex diffuse/normal that it was matched from
@@ -60,4 +59,7 @@ public:
    * @param NIFPath NIF that it was patched on
    */
   static void meshWarn(const std::wstring &MatchedPath, const std::wstring &NIFPath);
+
+  /// @brief print a summary of the already gathered warnings
+  static void printWarnings();
 };