+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fixed unicode character handling
 - Fixed exceptions when plugin patching is not enabled
 - PBR prefix check accounts for slot commands now too
+- Improved warning output for texture mismatches
 
 ## [0.7.1] - 2024-11-18